Efectos y etc

+++ client/includes/commands/command_sell_item.h
@@ -5,10 +5,10 @@
 class SellItemCommandClient: public ClientCommand {
 private:
     uint32_t npc_id;
-    uint16_t item_id;
+    uint8_t item_id;
     uint16_t quantity;
 
 public:
-    SellItemCommandClient(uint32_t npc_id, uint16_t item_id, uint16_t quantity);
+    SellItemCommandClient(uint32_t npc_id, uint8_t item_id, uint16_t quantity);
     void execute(ClientProtocol& protocol) const override;
 };
